backend/requirements.txt
# LangGraph 和 LangChain
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

# Web 框架
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# 环境变量管理
python-dotenv>=1.0.0

# 天文计算（可选，用于精确计算）
# ephem>=4.1.5

# 时区处理
pytz>=2023.3

# CORS 支持（已包含在 FastAPI 中，但列出以明确依赖）
# fastapi 已包含 CORSMiddleware

